refactor: Set appium as peer dependency
