+++ lib/gecko.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import os from 'os';
 import path from 'path';
-import { JWProxy, errors } from '@appium/base-driver';
-import { fs, util, system } from '@appium/support';
+import { JWProxy, errors } from 'appium/driver';
+import { fs, util, system } from 'appium/support';
 import { SubProcess } from 'teen_process';
 import { waitForCondition } from 'asyncbox';
 import { findAPortNotInUse } from 'portscanner';
